+++ package/CHANGELOG.md
@@ -1,11 +1,12 @@
 ## 0.4.0
 
-- Added an util to disable leaderboards
+- Added an util to disable leaderboards (for Thorn)
+- Added caching for image dominant color extraction (especially helps CybergrindSlideshow)
 
 ## 0.3.0
 
-- Added GameObject.SetActiveAnimated extension for easier UI panel animation
-- Added some small image processing and color utilities
+- Added GameObject.SetActiveAnimated extension for easier UI panel animation (for Taberry)
+- Added some small image processing and color utilities (for CybergrindSlideshow)
 
 ## 0.2.0